# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  controllers.Dentist:
    properties:
      birthday:
        type: string
      degree:
        type: integer
      expert:
        type: integer
      gender:
        type: integer
      nurse:
        type: integer
    type: object
...
  ent.Appointment:
    properties:
      datetime:
        description: Datetime holds the value of the "datetime" field.
        type: string
      detail:
        description: Detail holds the value of the "detail" field.
        type: string
      edges:
        $ref: '#/definitions/ent.AppointmentEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the AppointmentQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.AppointmentEdges:
    properties:
      dentist:
        $ref: '#/definitions/ent.Dentist'
        description: Dentist holds the value of the dentist edge.
        type: object
      patient:
        $ref: '#/definitions/ent.Patient'
        description: Patient holds the value of the patient edge.
        type: object
      room:
        $ref: '#/definitions/ent.Room'
        description: Room holds the value of the room edge.
        type: object
    type: object
  ent.Degree:
    properties:
      edges:
        $ref: '#/definitions/ent.DegreeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DegreeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.DegreeEdges:
    properties:
      dentists:
        description: Dentists holds the value of the dentists edge.
        items:
          $ref: '#/definitions/ent.Dentist'
        type: array
    type: object
  ent.DentalExpense:
    properties:
      added_time:
        description: AddedTime holds the value of the "added_time" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DentalExpenseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DentalExpenseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.DentalExpenseEdges:
    properties:
      medicalfile:
        $ref: '#/definitions/ent.Medicalfile'
        description: Medicalfile holds the value of the medicalfile edge.
        type: object
      nurse:
        $ref: '#/definitions/ent.Nurse'
        description: Nurse holds the value of the nurse edge.
        type: object
      pricetype:
        $ref: '#/definitions/ent.PriceType'
        description: Pricetype holds the value of the pricetype edge.
        type: object
    type: object
  ent.Dentist:
    properties:
      age:
        description: Age holds the value of the "age" field.
        type: integer
      birthday:
        description: Birthday holds the value of the "birthday" field.
        type: string
      cardid:
        description: Cardid holds the value of the "cardid" field.
        type: string
      edges:
        $ref: '#/definitions/ent.DentistEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DentistQuery when eager-loading is set.
        type: object
      email:
        description: Email holds the value of the "email" field.
        type: string
      experience:
        description: Experience holds the value of the "experience" field.
        type: string
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      password:
        description: Password holds the value of the "password" field.
        type: string
      tel:
        description: Tel holds the value of the "tel" field.
        type: string
    type: object
  ent.DentistEdges:
    properties:
      appointment:
        description: Appointment holds the value of the appointment edge.
        items:
          $ref: '#/definitions/ent.Appointment'
        type: array
      degree:
        $ref: '#/definitions/ent.Degree'
        description: Degree holds the value of the degree edge.
        type: object
      expert:
        $ref: '#/definitions/ent.Expert'
        description: Expert holds the value of the expert edge.
        type: object
      gender:
        $ref: '#/definitions/ent.Gender'
        description: Gender holds the value of the gender edge.
        type: object
      medicalfiles:
        description: Medicalfiles holds the value of the medicalfiles edge.
        items:
          $ref: '#/definitions/ent.Medicalfile'
        type: array
      nurse:
        $ref: '#/definitions/ent.Nurse'
        description: Nurse holds the value of the nurse edge.
        type: object
      queue:
        description: Queue holds the value of the queue edge.
        items:
          $ref: '#/definitions/ent.Queue'
        type: array
    type: object
  ent.Disease:
    properties:
      edges:
        $ref: '#/definitions/ent.DiseaseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the DiseaseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.DiseaseEdges:
    properties:
      patients:
        description: Patients holds the value of the patients edge.
        items:
          $ref: '#/definitions/ent.Patient'
        type: array
    type: object
  ent.Expert:
    properties:
      edges:
        $ref: '#/definitions/ent.ExpertEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the ExpertQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.ExpertEdges:
    properties:
      dentists:
        description: Dentists holds the value of the dentists edge.
        items:
          $ref: '#/definitions/ent.Dentist'
        type: array
    type: object
  ent.Gender:
    properties:
      edges:
        $ref: '#/definitions/ent.GenderEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the GenderQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.GenderEdges:
    properties:
      dentists:
        description: Dentists holds the value of the dentists edge.
        items:
          $ref: '#/definitions/ent.Dentist'
        type: array
      patients:
        description: Patients holds the value of the patients edge.
        items:
          $ref: '#/definitions/ent.Patient'
        type: array
    type: object
  ent.MedicalCare:
    properties:
      edges:
        $ref: '#/definitions/ent.MedicalCareEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MedicalCareQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.MedicalCareEdges:
    properties:
      medicalfiles:
        description: Medicalfiles holds the value of the medicalfiles edge.
        items:
          $ref: '#/definitions/ent.Medicalfile'
        type: array
      patients:
        description: Patients holds the value of the patients edge.
        items:
          $ref: '#/definitions/ent.Patient'
        type: array
    type: object
  ent.Medicalfile:
    properties:
      added_time:
        description: AddedTime holds the value of the "added_time" field.
        type: string
      detail:
        description: Detail holds the value of the "detail" field.
        type: string
      edges:
        $ref: '#/definitions/ent.MedicalfileEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the MedicalfileQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
    type: object
  ent.MedicalfileEdges:
    properties:
      dentalexpenses:
        description: Dentalexpenses holds the value of the dentalexpenses edge.
        items:
          $ref: '#/definitions/ent.DentalExpense'
        type: array
      dentist:
        $ref: '#/definitions/ent.Dentist'
        description: Dentist holds the value of the dentist edge.
        type: object
      medicalcare:
        $ref: '#/definitions/ent.MedicalCare'
        description: Medicalcare holds the value of the medicalcare edge.
        type: object
      nurse:
        $ref: '#/definitions/ent.Nurse'
        description: Nurse holds the value of the nurse edge.
        type: object
      patient:
        $ref: '#/definitions/ent.Patient'
        description: Patient holds the value of the patient edge.
        type: object
    type: object
  ent.Nurse:
    properties:
      edges:
        $ref: '#/definitions/ent.NurseEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the NurseQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      nurse_age:
        description: NurseAge holds the value of the "nurse_age" field.
        type: integer
      nurse_email:
        description: NurseEmail holds the value of the "nurse_email" field.
        type: string
      nurse_name:
        description: NurseName holds the value of the "nurse_name" field.
        type: string
      nurse_password:
        description: NursePassword holds the value of the "nurse_password" field.
        type: string
    type: object
  ent.NurseEdges:
    properties:
      dentalexpenses:
        description: Dentalexpenses holds the value of the dentalexpenses edge.
        items:
          $ref: '#/definitions/ent.DentalExpense'
        type: array
      dentists:
        description: Dentists holds the value of the dentists edge.
        items:
          $ref: '#/definitions/ent.Dentist'
        type: array
      medicalfiles:
        description: Medicalfiles holds the value of the medicalfiles edge.
        items:
          $ref: '#/definitions/ent.Medicalfile'
        type: array
      patients:
        description: Patients holds the value of the patients edge.
        items:
          $ref: '#/definitions/ent.Patient'
        type: array
      queue:
        description: Queue holds the value of the queue edge.
        items:
          $ref: '#/definitions/ent.Queue'
        type: array
    type: object
  ent.Patient:
    properties:
      age:
        description: Age holds the value of the "age" field.
        type: integer
      birthday:
        description: Birthday holds the value of the "birthday" field.
        type: string
      cardID:
        description: CardID holds the value of the "cardID" field.
        type: string
      edges:
        $ref: '#/definitions/ent.PatientEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PatientQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
      patient_ID:
        description: PatientID holds the value of the "patient_ID" field.
        type: string
      tel:
        description: Tel holds the value of the "tel" field.
        type: string
    type: object
  ent.PatientEdges:
    properties:
      appointment:
        description: Appointment holds the value of the appointment edge.
        items:
          $ref: '#/definitions/ent.Appointment'
        type: array
      disease:
        $ref: '#/definitions/ent.Disease'
        description: Disease holds the value of the disease edge.
        type: object
      gender:
        $ref: '#/definitions/ent.Gender'
        description: Gender holds the value of the gender edge.
        type: object
      medicalcare:
        $ref: '#/definitions/ent.MedicalCare'
        description: Medicalcare holds the value of the medicalcare edge.
        type: object
      medicalfiles:
        description: Medicalfiles holds the value of the medicalfiles edge.
        items:
          $ref: '#/definitions/ent.Medicalfile'
        type: array
      nurse:
        $ref: '#/definitions/ent.Nurse'
        description: Nurse holds the value of the nurse edge.
        type: object
      queue:
        description: Queue holds the value of the queue edge.
        items:
          $ref: '#/definitions/ent.Queue'
        type: array
    type: object
  ent.PriceType:
    properties:
      edges:
        $ref: '#/definitions/ent.PriceTypeEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the PriceTypeQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.PriceTypeEdges:
    properties:
      dentalexpenses:
        description: Dentalexpenses holds the value of the dentalexpenses edge.
        items:
          $ref: '#/definitions/ent.DentalExpense'
        type: array
    type: object
  ent.Queue:
    properties:
      dental:
        description: Dental holds the value of the "dental" field.
        type: string
      edges:
        $ref: '#/definitions/ent.QueueEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the QueueQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      queue_time:
        description: QueueTime holds the value of the "queue_time" field.
        type: string
    type: object
  ent.QueueEdges:
    properties:
      dentist:
        $ref: '#/definitions/ent.Dentist'
        description: Dentist holds the value of the dentist edge.
        type: object
      nurse:
        $ref: '#/definitions/ent.Nurse'
        description: Nurse holds the value of the nurse edge.
        type: object
      patient:
        $ref: '#/definitions/ent.Patient'
        description: Patient holds the value of the patient edge.
        type: object
    type: object
  ent.Room:
    properties:
      edges:
        $ref: '#/definitions/ent.RoomEdges'
        description: |-
          Edges holds the relations/edges for other nodes in the graph.
          The values are being populated by the RoomQuery when eager-loading is set.
        type: object
      id:
        description: ID of the ent.
        type: integer
      name:
        description: Name holds the value of the "name" field.
        type: string
    type: object
  ent.RoomEdges:
    properties:
      appointment:
        description: Appointment holds the value of the appointment edge.
        items:
          $ref: '#/definitions/ent.Appointment'
        type: array
    type: object
  gin.H:
    additionalProperties: true
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for SUT SE 2563
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SUT SA Example API
  version: "1.0"
paths:
...
  /degrees:
    get:
      description: list degree entities
      operationId: list-degree
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Degree'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List degree entities
    post:
      consumes:
      - application/json
      description: Create degree
      operationId: create-degree
      parameters:
      - description: ' Degree entity'
        in: body
        name: degree
        required: true
        schema:
          $ref: '#/definitions/ent.Degree'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Degree'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create degree
  /degrees/{id}:
    get:
      description: get degree by ID
      operationId: get-degree
      parameters:
      - description: Degree ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Degree'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a Degree entity by ID
...
  /dentists:
    get:
      description: list dentist entities
      operationId: list-dentist
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Dentist'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List dentist entities
    post:
      consumes:
      - application/json
      description: Create dentist
      operationId: create-dentist
      parameters:
      - description: Dentist entity
        in: body
        name: dentist
        required: true
        schema:
          $ref: '#/definitions/controllers.Dentist'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.Dentist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create dentist
  /dentists/{id}:
    get:
      description: get dentist by ID
      operationId: get-dentist
      parameters:
      - description: Dentist ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Dentist'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a dentist entity by ID
...
  /genders:
    get:
      description: list gender entities
      operationId: list-gender
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/ent.Gender'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: List gender entities
    post:
      consumes:
      - application/json
      description: Create gender
      operationId: create-gender
      parameters:
      - description: Gender entity
        in: body
        name: gender
        required: true
        schema:
          $ref: '#/definitions/ent.Gender'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Gender'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Create gender
  /genders/{id}:
    get:
      description: get gender by ID
      operationId: get-gender
      parameters:
      - description: Gender ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ent.Gender'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gin.H'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gin.H'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gin.H'
      summary: Get a gender entity by ID
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
  OAuth2AccessCode:
    authorizationUrl: https://example.com/oauth/authorize
    flow: accessCode
    scopes:
      admin: ' Grants read and write access to administrative information'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Application:
    authorizationUrl: ""
    flow: application
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
  OAuth2Implicit:
    authorizationUrl: https://example.com/oauth/authorize
    flow: implicit
    scopes:
      admin: ' Grants read and write access to administrative information'
      write: ' Grants write access'
    type: oauth2
  OAuth2Password:
    authorizationUrl: ""
    flow: password
    scopes:
      admin: ' Grants read and write access to administrative information'
      read: ' Grants read access'
      write: ' Grants write access'
    tokenUrl: https://example.com/oauth/token
    type: oauth2
swagger: "2.0"